moved tests into the lbrynet package
