+++ lbrynet/tests/functional/test_reflector.py
@@ -19,8 +19,8 @@ from lbrynet.lbry_file.client import EncryptedFileOptions
 from lbrynet.file_manager import EncryptedFileCreator
 from lbrynet.file_manager import EncryptedFileManager
 
-from tests import mocks
-from tests.util import mk_db_and_blob_dir, rm_db_and_blob_dir
+from lbrynet.tests import mocks
+from lbrynet.tests.util import mk_db_and_blob_dir, rm_db_and_blob_dir
 
 class TestReflector(unittest.TestCase):
     def setUp(self):
