+++ lbrynet/tests/unit/core/test_log_support.py
