+++ lbrynet/tests/unit/lbrynet_daemon/test_DaemonCLI.py
